Move implementation to @cosmwasm/stargate
