+++ packages/cosmwasm-stargate/src/index.ts
@@ -1,5 +1,6 @@
 export { Code, CodeDetails, Contract, ContractCodeHistoryEntry, CosmWasmClient } from "./cosmwasmclient";
 export { fromBinary, toBinary } from "./encoding";
+export { _instantiate2AddressIntermediate, instantiate2Address } from "./instantiate2_address";
 export {
   cosmWasmTypes,
   createWasmAminoConverters,
